requirements.txt
MySQL-python==1.2.5
Django==1.6.5
django-mptt==0.6.1

